initial commit for proxy server
The proxy server keeps the api key protected. I've also add a json file
listing all of the agencies in the US. This list will be used for
finding agencies within a range of a requested latitude and longitude
so request for crime data can be made for those selected agencies.